Added sealing interface
